Recoded play_sound in a seperate script file + set it as autoload
Had to recode the play_sound function from scratch due to numerous issues with how it was implemented, all audio nodes would be queue_free'd if an audio node finished playing it's sound which resulted in every other audio node that hasn't finishing their sound to be queue_free'd prematurely, resulting in an audible sudden cut in the sound or game crash due to non-existing audio nodes if spammed too often.
This recode works as intended, I have tested and debugged it with the Debugger tool in Godot.