Updated offset, collisionshape,  node-structure, shadows
I have reset the tree's position that had been changed by mistake beforehand. The shadow of the tree is now more subtle and wider. I have changed the node structure so that the collisionshape is visible when in the editor. Lastly i have changed the collisionshape from a Rectangleshape to a Capsule shape so the player won't get stuck.