Pine tree sprite change and organized tree folder structure (#31)
* Pine tree sprite change and organized tree folder structure

I have changed the pine_tree.tscn sprite to be an animated pixel tree. I have also made the tree bigger and organized the tree folder into two folders, Assets and Scenes.

* Updated offset, collisionshape,  node-structure, shadows

I have reset the tree's position that had been changed by mistake beforehand. The shadow of the tree is now more subtle and wider. I have changed the node structure so that the collisionshape is visible when in the editor. Lastly i have changed the collisionshape from a Rectangleshape to a Capsule shape so the player won't get stuck.