Changed Spider Properties (#40)
I have set the default speed to 300 to match  the other mobs. (This can of course be changed later if we add special properties to the different mobs). I have also set the motion mode to floating to stop the spider from sticking. I have also set the Collisionlayer to level 2 and the mask to level 1 and 2.