Various mob changes (#43)
* Moved all the mobs into a "Mobs" folder

I have moved all the mob scenes and assets into their separate folder inside a new "Mobs" folder. This is purely for organizational purposes.

* Exported ghost Speed variable and changed to standard mob speed

I have made an exported variable that replaces the constant speed of 350 that was set in ghost.gd. I have also changed the speed to the standard mob speed of 300.

* Corrected the preload adress in mob_spawning.tscn

Corrected the preload adress in mob_spawning.tscn.